confirmacion delete

+++ src/componentes/DeleteCharacter.jsx
@@ -2,20 +2,23 @@ import axios from "axios"
 import { useNavigate } from "react-router-dom"
 
 const DeleteCharacter = ({ id, onDelete }) => {
-const navigate = useNavigate()
+  const navigate = useNavigate()
 
-const deleteCharacter = async () => {
-await axios.delete(`http://localhost:8080/characters/${id}`)
+  const deleteCharacter = async () => {
+    const confirmed = window.confirm("¿Quieres eliminar este personaje?")
 
-    onDelete()
-    navigate("/")
-}
+    if (confirmed) {
+      await axios.delete(`http://localhost:8080/characters/${id}`)
+      onDelete()
+      navigate("/")
+    }
+  }
 
-return (
+  return (
     <div className="delete">
-    <button onClick={deleteCharacter}>Eliminar personaje</button>
+      <button onClick={deleteCharacter}>Eliminar personaje</button>
     </div>
-)
+  )
 }
 
 export default DeleteCharacter
